reactualisation vector pour testeur (suppression fct advance, compilation avec warning

+++ vector.hpp
@@ -474,7 +474,7 @@ namespace ft
 	{
 		if (lhs.size() != rhs.size())
 			return (false);
-		for (int i = 0; i < lhs.size(); i++)
+		for (size_t i = 0; i < lhs.size(); i++)
 		{
 			if (lhs[i] != rhs[i])
 				return (false);
